refactor(web-app): update import paths for landing components
- Adjusted import paths for various landing components to reflect the new directory structure.
- This change improves organization and consistency across the codebase.
